# docs/openapi.yaml
openapi: 3.0.3
info:
  title: GreenMetrics API (MVP)
  version: 0.1.0
  description: ESG reporting platform API for EU SMEs
servers:
  - url: https://app.greenmetrics.example/api/v1
    description: Production server
  - url: http://localhost:3000/api/v1
    description: Development server
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth session cookie
  schemas:
...
    Report:
      type: object
      required: [id, organizationId, periodStart, periodEnd, status, createdAt]
      properties:
        id: { type: string, format: uuid }
        organizationId: { type: string, format: uuid }
        periodStart: { type: string, format: date }
        periodEnd: { type: string, format: date }
        status: { type: string, enum: [DRAFT, COMPLETED] }
        createdBy: { type: string, format: uuid }
        createdAt: { type: string, format: date-time }
        pdfUrl: { type: string, format: uri, nullable: true }
      example:
        id: '456e7890-e89b-12d3-a456-426614174001'
        organizationId: '123e4567-e89b-12d3-a456-426614174000'
        periodStart: '2024-01-01'
        periodEnd: '2024-12-31'
        status: 'COMPLETED'
        createdBy: '789e0123-e89b-12d3-a456-426614174002'
        createdAt: '2025-08-15T10:30:00Z'
        pdfUrl: 'https://storage.example.com/reports/456e7890.pdf'
...
security:
  - sessionAuth: []
paths:
...
  /reports:
    post:
      summary: Create a report (draft)
      tags: [Reports]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [organizationId, periodStart, periodEnd]
              properties:
                organizationId: { type: string, format: uuid }
                periodStart: { type: string, format: date }
                periodEnd: { type: string, format: date }
            example:
              organizationId: '123e4567-e89b-12d3-a456-426614174000'
              periodStart: '2024-01-01'
              periodEnd: '2024-12-31'
      responses:
        '201': 
          description: Report created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Report'
    get:
      summary: List reports for an organization
      tags: [Reports]
      parameters:
        - in: query
          name: organizationId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': 
          description: List of reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Report'
...